requirements.txt
Flask==2.3.3
Flask-CORS==4.0.1
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.0
matplotlib==3.8.0
typing-extensions==4.9.0
six==1.16.0